+++ app/src/main/java/com/asfoundation/wallet/ui/iab/IabView.java
@@ -1,5 +1,7 @@
 package com.asfoundation.wallet.ui.iab;
 
+import java.math.BigDecimal;
+
 /**
  * Created by franciscocalado on 20/07/2018.
  */
@@ -10,7 +12,7 @@ public interface IabView {
 
   void close();
 
-  void setup(double amount, Boolean canBuy);
+  void setup(BigDecimal amount, Boolean canBuy);
 
   void navigateToCreditCardAuthorization();
 }
